+++ Assets/_Projects/Scripts/UI/PopupIapDetail/PopupIapDetail.cs
@@ -1,6 +1,4 @@
 using Sirenix.OdinInspector;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -17,7 +15,9 @@ public class PopupIapDetail : DraftUtils.DraftMonoBehaviour
     private void Start()
     {
         popup.closeButton.OnClickAction = popup.HideWithAnimation;
+        buyButton.onClick.AddListener(ClickBuyButton);
     }
+
     public void SetData(MultipleIAPData data)
     {
         _data = data;
@@ -27,5 +27,39 @@ public class PopupIapDetail : DraftUtils.DraftMonoBehaviour
     private void SetMultipleIAPDataView()
     {
         multipleIAPDataView.SetData(_data);
+    }
+
+    private void ClickBuyButton()
+    {
+        if (_data == null)
+        {
+            return;
+        }
+
+        string productId = string.IsNullOrEmpty(_data.productId)
+            ? _data.itemType.ToString()
+            : _data.productId;
+
+        DraftUtils.IAP.IAPManager.Instance.Purchase(productId, result =>
+        {
+            if (!result.IsSuccess)
+            {
+                GameAnalytics.LogPurchaseEvent(
+                    GameAnalytics.IapPurchaseFail,
+                    productId,
+                    result.FailureReason.ToString());
+                return;
+            }
+
+            GameAnalytics.LogPurchaseEvent(GameAnalytics.IapPurchaseSuccess, productId);
+
+            var rewards = MultipleIAPDataExtensions.GetRewards(_data);
+            if (rewards.Count > 0)
+            {
+                DataManager.Instance.Reward(rewards);
+            }
+
+            PopupManager.Instance.ShowPopupMultipleIapReward(_data);
+        });
     }
 }